+++ packages/cli/src/export/__tests__/writeAssetsToCsv.test.ts
@@ -1,5 +1,5 @@
 import { vol } from 'memfs';
-import { v4 as uuid } from 'uuid';
+import { randomUUID as uuid } from 'crypto';
 
 import { groupJsonAssetsByType } from '../groupJsonAssetsByType';
 import { TEST_STORAGE_LOCATION } from '../../__tests__/utils';